yfinance_service/requirements.txt
yfinance
psycopg2-binary
python-dotenv
# requests